gizmo/requirements.txt
flask==3.1.2
flask-cors==6.0.2
pymongo==4.7.0
python-dotenv==1.0.1
llama-cpp-python==0.2.90
dnspython==2.8.0
Flask-JWT-Extended==4.7.1
Flask-Bcrypt==1.0.1
